start auto mapper

+++ FrogExhibitionBLL/Services/VoteService.cs
@@ -33,28 +33,35 @@ namespace FrogExhibitionBLL.Services
         {
             try
             {
-                var currentUserId = await _userProvider.GetUserIdAsync();
-                var userVotesOnExhibitionCount = _unitOfWork.Votes
-                    .GetAllQueryable()
-                    .Where(v => v.ApplicationUserId == currentUserId 
-                    && v.FrogOnExhibitionId == vote.FrogOnExhibitionId)
-                    .Count();
-                if (userVotesOnExhibitionCount >= 3)
+                if (await _unitOfWork.FrogOnExhibitions.EntityExistsAsync(vote.FrogOnExhibitionId))
                 {
-                    throw new DbUpdateException("This user has cast all of his available votes on this exebiton");
+                    var currentUserId = await _userProvider.GetUserIdAsync();
+                    var userVotesOnExhibitionCount = _unitOfWork.Votes
+                        .GetAllQueryable()
+                        .Where(v => v.ApplicationUserId == currentUserId
+                        && v.FrogOnExhibitionId == vote.FrogOnExhibitionId)
+                        .Count();
+                    if (userVotesOnExhibitionCount >= 3)
+                    {
+                        throw new DbUpdateException("This user has cast all of his available votes on this exebiton");
+                    }
+                    try
+                    {
+                        var mappedVote = _mapper.Map<Vote>(vote);
+                        mappedVote.ApplicationUserId = currentUserId;
+                        await _unitOfWork.Votes.CreateAsync(mappedVote);
+                        _logger.LogInformation("Vote Created");
+                        await _unitOfWork.SaveAsync();
+                        return mappedVote.Id;
+                    }
+                    catch (DbUpdateException ex)
+                    {
+                        throw new DbUpdateException("You can't vote for the same frog on exebition twice");
+                    }
                 }
-                try
+                else
                 {
-                    var mappedVote = _mapper.Map<Vote>(vote);
-                    mappedVote.ApplicationUserId = currentUserId;
-                    await _unitOfWork.Votes.CreateAsync(mappedVote);
-                    _logger.LogInformation("Vote Created");
-                    await _unitOfWork.SaveAsync();
-                    return mappedVote.Id;
-                }
-                catch(DbUpdateException ex)
-                {
-                    throw new DbUpdateException("You can't vote for the same frog on exebition twice");
+                    throw new NotFoundException("Can't find frog at exhibition");
                 }
             }
             catch (Exception ex)
@@ -86,19 +93,26 @@ namespace FrogExhibitionBLL.Services
             {
                 if (await _unitOfWork.Votes.EntityExistsAsync(vote.Id))
                 {
-                    var currentUserId = await _userProvider.GetUserIdAsync();
-                    var userVotesOnExhibitionCount = _unitOfWork.Votes.GetAllQueryable()
-                        .Where(v => v.ApplicationUserId == currentUserId 
-                        && v.FrogOnExhibitionId == vote.FrogOnExhibitionId)
-                        .Count();
-                    if (userVotesOnExhibitionCount >= 3)
+                    if (await _unitOfWork.FrogOnExhibitions.EntityExistsAsync(vote.FrogOnExhibitionId))
                     {
-                        throw new DbUpdateException("This user has cast all of his available votes on this exebiton");
+                        var currentUserId = await _userProvider.GetUserIdAsync();
+                        var userVotesOnExhibitionCount = _unitOfWork.Votes.GetAllQueryable()
+                            .Where(v => v.ApplicationUserId == currentUserId
+                            && v.FrogOnExhibitionId == vote.FrogOnExhibitionId)
+                            .Count();
+                        if (userVotesOnExhibitionCount >= 3)
+                        {
+                            throw new DbUpdateException("This user has cast all of his available votes on this exebiton");
+                        }
+                        var mappedVote = _mapper.Map<Vote>(vote);
+                        mappedVote.ApplicationUserId = currentUserId;
+                        await _unitOfWork.Votes.UpdateAsync(mappedVote);
+                        await _unitOfWork.SaveAsync();
                     }
-                    var mappedVote = _mapper.Map<Vote>(vote);
-                    mappedVote.ApplicationUserId = currentUserId;
-                    await _unitOfWork.Votes.UpdateAsync(mappedVote);
-                    await _unitOfWork.SaveAsync();
+                    else
+                    {
+                        throw new NotFoundException("Frog on exhibition not found");
+                    }
                 }
                 else
                 {
@@ -126,13 +140,23 @@ namespace FrogExhibitionBLL.Services
 
         public async Task DeleteVoteAsync(Guid id)
         {
+
+            var currentUserId = await _userProvider.GetUserIdAsync();
             var vote = await _unitOfWork.Votes.GetAsync(id);
             if (vote == null)
             {
                 throw new NotFoundException("Entity not found");
             }
-            await _unitOfWork.Votes.DeleteAsync(vote.Id);
-            await _unitOfWork.SaveAsync();
+            if (vote.ApplicationUserId == currentUserId)
+            {
+                await _unitOfWork.Votes.DeleteAsync(vote.Id);
+                await _unitOfWork.SaveAsync();
+            }
+            else
+            {
+                throw new ForbidException("You can't delete other users votes");
+            }
+            
         }
     }
 }
